+++ lib/tests/templates.test.ts
@@ -14,6 +14,7 @@ import {
   deleteTemplate,
   getTemplateSubmissionTypeByID,
   onlyIncludePublicProperties,
+  updateIsPublishedForTemplate,
 } from "../templates";
 
 import { BetterOmit, FormRecord } from "@lib/types";
@@ -24,7 +25,7 @@ import formConfiguration from "@jestFixtures/cdsIntakeTestForm.json";
 import v8 from "v8";
 import { Prisma } from "@prisma/client";
 import { AccessControlError, createAbility } from "@lib/privileges";
-import { Base, getUserPrivileges, ManageForms } from "__utils__/permissions";
+import { Base, getUserPrivileges, ManageForms, PublishForms } from "__utils__/permissions";
 
 const redis = new Redis();
 
@@ -61,7 +62,7 @@ describe("Template CRUD functions", () => {
     const newTemplate = await createTemplate(
       ability,
       "1",
-      formConfiguration as BetterOmit<FormRecord, "id" | "bearerToken">
+      formConfiguration as BetterOmit<FormRecord, "id" | "isPublished">
     );
 
     expect(prismaMock.template.create).toHaveBeenCalledWith({
@@ -135,6 +136,7 @@ describe("Template CRUD functions", () => {
       select: {
         id: true,
         jsonConfig: true,
+        isPublished: true,
       },
     });
   });
@@ -163,6 +165,7 @@ describe("Template CRUD functions", () => {
       select: {
         id: true,
         jsonConfig: true,
+        isPublished: true,
       },
     });
   });
@@ -182,6 +185,7 @@ describe("Template CRUD functions", () => {
       select: {
         id: true,
         jsonConfig: true,
+        isPublished: true,
         ttl: true,
       },
     });
@@ -231,12 +235,12 @@ describe("Template CRUD functions", () => {
     });
 
     const updatedFormConfig = structuredClone(
-      formConfiguration as BetterOmit<FormRecord, "id" | "bearerToken">
+      formConfiguration as BetterOmit<FormRecord, "id" | "isPublished">
     );
-    updatedFormConfig.publishingStatus = true;
 
     (prismaMock.template.update as jest.MockedFunction<any>).mockResolvedValue({
       id: "formtestID",
+      isPublished: true,
       jsonConfig: updatedFormConfig,
     });
 
@@ -252,12 +256,88 @@ describe("Template CRUD functions", () => {
       select: {
         id: true,
         jsonConfig: true,
+        isPublished: true,
       },
     });
 
     expect(updatedTemplate).toEqual({
       id: "formtestID",
+      isPublished: true,
       ...updatedFormConfig,
+    });
+  });
+
+  it("Update `isPublished` on a specific form", async () => {
+    const ability = createAbility(getUserPrivileges(PublishForms, { user: { id: "1" } }));
+
+    (prismaMock.template.findUnique as jest.MockedFunction<any>).mockResolvedValue({
+      id: "formtestID",
+      jsonConfig: formConfiguration,
+      users: [{ id: "1" }],
+    });
+
+    (prismaMock.template.update as jest.MockedFunction<any>).mockResolvedValue({
+      id: "formtestID",
+      jsonConfig: formConfiguration,
+      isPublished: true,
+    });
+
+    const updatedTemplate = await updateIsPublishedForTemplate(ability, "formtestID", true);
+
+    expect(prismaMock.template.update).toHaveBeenCalledWith({
+      where: {
+        id: "formtestID",
+      },
+      data: {
+        isPublished: true,
+      },
+      select: {
+        id: true,
+        jsonConfig: true,
+        isPublished: true,
+      },
+    });
+
+    expect(updatedTemplate).toEqual({
+      id: "formtestID",
+      isPublished: true,
+      ...formConfiguration,
+    });
+  });
+
+  /**
+   * Can be deleted once we think that the progressive migration is not needed anymore and the code in lib/templates.ts has been removed
+   */
+  it("Update `isPublished` on a specific form should handle soft migration with existing `publishingStatus`", async () => {
+    const ability = createAbility(getUserPrivileges(PublishForms, { user: { id: "1" } }));
+
+    (prismaMock.template.findUnique as jest.MockedFunction<any>).mockResolvedValue({
+      id: "formtestID",
+      jsonConfig: { ...formConfiguration, publishingStatus: true },
+      users: [{ id: "1" }],
+    });
+
+    (prismaMock.template.update as jest.MockedFunction<any>).mockResolvedValue({
+      id: "formtestID",
+      jsonConfig: formConfiguration,
+      isPublished: true,
+    });
+
+    await updateIsPublishedForTemplate(ability, "formtestID", false);
+
+    expect(prismaMock.template.update).toHaveBeenCalledWith({
+      where: {
+        id: "formtestID",
+      },
+      data: {
+        isPublished: false,
+        jsonConfig: formConfiguration as unknown as Prisma.JsonObject,
+      },
+      select: {
+        id: true,
+        jsonConfig: true,
+        isPublished: true,
+      },
     });
   });
 
@@ -288,6 +368,7 @@ describe("Template CRUD functions", () => {
         select: {
           id: true,
           jsonConfig: true,
+          isPublished: true,
         },
       })
     );
@@ -302,6 +383,7 @@ describe("Template CRUD functions", () => {
     const formRecord = {
       id: "testID",
       ...formConfiguration,
+      isPublished: false,
     };
 
     const publicFormRecord = onlyIncludePublicProperties(formRecord as FormRecord);
@@ -325,7 +407,7 @@ describe("Template CRUD functions", () => {
       await createTemplate(
         ability,
         "1",
-        formConfiguration as BetterOmit<FormRecord, "id" | "bearerToken">
+        formConfiguration as BetterOmit<FormRecord, "id" | "isPublished">
       );
     }).rejects.toThrowError(new AccessControlError(`Access Control Forbidden Action`));
 
@@ -337,8 +419,12 @@ describe("Template CRUD functions", () => {
       await updateTemplate(
         ability,
         "test1",
-        structuredClone(formConfiguration as BetterOmit<FormRecord, "id" | "bearerToken">)
+        structuredClone(formConfiguration as BetterOmit<FormRecord, "id" | "isPublished">)
       );
+    }).rejects.toThrowError(new AccessControlError(`Access Control Forbidden Action`));
+
+    await expect(async () => {
+      await updateIsPublishedForTemplate(ability, "formtestID", true);
     }).rejects.toThrowError(new AccessControlError(`Access Control Forbidden Action`));
 
     await expect(async () => {
@@ -359,8 +445,12 @@ describe("Template CRUD functions", () => {
       await updateTemplate(
         ability,
         "test1",
-        structuredClone(formConfiguration as BetterOmit<FormRecord, "id" | "bearerToken">)
+        structuredClone(formConfiguration as BetterOmit<FormRecord, "id" | "isPublished">)
       );
+    }).rejects.toThrowError(new AccessControlError(`Access Control Forbidden Action`));
+
+    await expect(async () => {
+      await updateIsPublishedForTemplate(ability, "formtestID", true);
     }).rejects.toThrowError(new AccessControlError(`Access Control Forbidden Action`));
 
     await expect(async () => {