+++ prisma/seeds/fixtures/privileges.ts
@@ -19,7 +19,7 @@ const Base: PrivilegeSeed = {
       subject: "FormRecord",
       conditions: { users: { $elemMatch: { id: "${user.id}" } } },
     },
-    { action: "update", subject: "FormRecord", fields: ["publishingStatus"], inverted: true },
+    { action: "update", subject: "FormRecord", fields: ["isPublished"], inverted: true },
   ],
   priority: 0,
 };
@@ -33,7 +33,7 @@ const PublishForms: PrivilegeSeed = {
     {
       action: ["update"],
       subject: "FormRecord",
-      fields: ["publishingStatus"],
+      fields: ["isPublished"],
       conditions: { users: { $elemMatch: { id: "${user.id}" } } },
     },
   ],